+++ CumulusMX/StationSettings.cs
@@ -11,6 +11,8 @@ using EmbedIO;
 using ServiceStack;
 using ServiceStack.Text;
 
+using static System.Collections.Specialized.BitVector32;
+
 
 namespace CumulusMX
 {
@@ -1267,6 +1269,10 @@ namespace CumulusMX
 								cumulus.Limit.PressLow = ConvertUnits.UserPressToKpa(cumulus.Limit.PressLow);
 								break;
 						}
+
+						cumulus.Limit.StationPressHigh = ConvertUnits.PressMBToUser(MeteoLib.SeaLevelToStation(ConvertUnits.UserPressToHpa(cumulus.Limit.PressHigh), station.AltitudeM(cumulus.Altitude)));
+						cumulus.Limit.StationPressLow = ConvertUnits.PressMBToUser(MeteoLib.SeaLevelToStation(ConvertUnits.UserPressToHpa(cumulus.Limit.PressLow), station.AltitudeM(cumulus.Altitude)));
+
 						cumulus.Units.Press = settings.general.units.pressure;
 						cumulus.ChangePressureUnits();
 						settings.general.units.advanced.pressdp = cumulus.PressDPlaceDefaults[cumulus.Units.Press];